+++ src/main/components/ArrangeView/ArrangeViewCanvas/ArrangeViewCanvas.tsx
@@ -3,9 +3,9 @@ import { observer } from "mobx-react-lite"
 import { useMemo, useRef, VFC } from "react"
 import { matrixFromTranslation } from "../../../helpers/matrix"
 import { useStores } from "../../../hooks/useStores"
-import { Beats } from "../../GLSurface/common/Beats"
-import { Cursor } from "../../GLSurface/common/Cursor"
-import { Selection } from "../../GLSurface/common/Selection"
+import { Beats } from "../../GLNodes/Beats"
+import { Cursor } from "../../GLNodes/Cursor"
+import { Selection } from "../../GLNodes/Selection"
 import { Lines } from "./Lines"
 import { Notes } from "./Notes"
