+++ src/main/components/GLNodes/Selection.tsx
@@ -3,8 +3,8 @@ import { BorderedRectangles } from "@ryohey/webgl-react"
 import Color from "color"
 import { vec4 } from "gl-matrix"
 import { VFC } from "react"
-import { IRect } from "../../../../common/geometry"
-import { colorToVec4 } from "../../../gl/color"
+import { IRect } from "../../../common/geometry"
+import { colorToVec4 } from "../../gl/color"
 
 export const Selection: VFC<{ rect: IRect | null; zIndex: number }> = ({
   rect,
