+++ src/main/components/PianoRoll/PianoRollCanvas/PianoRollCanvas.tsx
@@ -11,9 +11,9 @@ import {
 import { matrixFromTranslation } from "../../../helpers/matrix"
 import { useContextMenu } from "../../../hooks/useContextMenu"
 import { useStores } from "../../../hooks/useStores"
-import { Beats } from "../../GLSurface/common/Beats"
-import { Cursor } from "../../GLSurface/common/Cursor"
-import { Selection } from "../../GLSurface/common/Selection"
+import { Beats } from "../../GLNodes/Beats"
+import { Cursor } from "../../GLNodes/Cursor"
+import { Selection } from "../../GLNodes/Selection"
 import NoteMouseHandler from "../MouseHandler/NoteMouseHandler"
 import { PianoRollStageProps } from "../PianoRollStage"
 import { PianoSelectionContextMenu } from "../PianoSelectionContextMenu"
