+++ src/main/components/TempoGraph/TempoGraphCanvas/TempoGraphCanvas.tsx
@@ -9,9 +9,9 @@ import {
 import { changeTempo } from "../../../actions"
 import { matrixFromTranslation } from "../../../helpers/matrix"
 import { useStores } from "../../../hooks/useStores"
-import { Beats } from "../../GLSurface/common/Beats"
-import { Cursor } from "../../GLSurface/common/Cursor"
-import { Selection } from "../../GLSurface/common/Selection"
+import { Beats } from "../../GLNodes/Beats"
+import { Cursor } from "../../GLNodes/Cursor"
+import { Selection } from "../../GLNodes/Selection"
 import { handleCreateSelectionDrag } from "../MouseHandler/handleCreateSelectionDrag"
 import { handlePencilMouseDown } from "../MouseHandler/handlePencilMouseDown"
 import { handleSelectionDragEvents } from "../MouseHandler/handleSelectionDragEvents"
